backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
playwright>=1.40.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dateutil>=2.8.2
slowapi>=0.1.9
html-sanitizer>=2.4.0
redis>=4.5.0
numpy>=1.24.0
